+++ magda-typescript-common/package.json
@@ -9,6 +9,7 @@
     "watch": "tsc -p tsconfig-build.json --watch",
     "generate": "generate-registry-typescript ./src/generated/registry",
     "test": "nyc mocha --compilers ts:ts-node/register,tsx:ts-node/register --require tsconfig-paths/register \"src/test/**/*.spec.ts\"",
+    "create_esri_jwt": "nyc mocha --compilers ts:ts-node/register,tsx:ts-node/register --require tsconfig-paths/register \"src/test/**/buildJwtForRegistryEsriOpaTest.ts\"",
     "dev": "yarn run watch"
   },
   "devDependencies": {
