+++ magda-typescript-common/src/test/session/buildJwt.spec.ts
@@ -50,43 +50,4 @@ describe("Get authz claim from jwt token", () => {
         const actual = getUserSession(req, argv.jwtSecret).valueOr({});
         expect(actual.session.esriGroups).to.be.deep.equal(groups);
     });
-
-    it("should create jwt token", function() {
-        const sessions = [
-            {
-                session: {
-                    esriGroups: [
-                        "Dep. A",
-                        "Branch A, Dep. A",
-                        "Branch B, Dep. A",
-                        "Section C, Branch B, Dep. A"
-                    ]
-                }
-            },
-            { session: { esriGroups: ["Branch A, Dep. A"] } },
-            {
-                session: {
-                    esriGroups: [
-                        "Branch B, Dep. A",
-                        "Section C, Branch B, Dep. A"
-                    ]
-                }
-            },
-            { session: { esriGroups: ["Section C, Branch B, Dep. A"] } }
-        ];
-
-        const userIds = [
-            "00000000-0000-1000-0000-000000000000",
-            "00000000-0000-1000-0001-000000000000",
-            "00000000-0000-1000-0002-000000000000",
-            "00000000-0000-1000-0003-000000000000"
-        ];
-
-        for (var i = 0; i < userIds.length; i++) {
-            let userId = userIds[i];
-            let session = sessions[i];
-            let jwtToken = buildJwt(argv.jwtSecret, userId, session);
-            console.log(`i: ${i}; \njwt: ${jwtToken}`);
-        }
-    });
 });
